api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
yfinance==0.2.38
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
databento==0.71.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9